VectorHRS/requirements.txt
Django==3.2.4
drf-yasg==1.20.0
mysqlclient==2.1.0  # Required for MySQL database backend
django-ckeditor==6.1.0
django-bleach==0.6.0
django-extensions==3.1.3
django-filter>=2.4.0
django-modelcluster>=5.5.1
djangorestframework==3.12.4
django-seed>=0.1.6
django-taggit>=1.5.1
gunicorn>=20.1.0
Pillow==8.4.0